refactor(chat): organize dir
