+++ front-end/src/features/chat/hooks/use-websocket-connection.ts
