+++ front-end/src/features/chat/services/connection-service.ts
